requirement.txt
absl-py==1.0.0
aiohttp==3.7.4
aiohttp-cors==0.7.0
aioredis==1.3.1
aiosignal==1.3.1
argcomplete==2.0.0
async-timeout==3.0.1
attrs==21.4.0
Automat==22.10.0
beautifulsoup4==4.10.0
black==21.12b0
blessings==1.7
boto==2.49.0
braceexpand==0.1.7
buildtools==1.0.6
cachetools==4.2.4
causal-conv1d==1.1.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.10
click==8.0.3
cloudpickle==2.1.0
colorama==0.4.4
colorful==0.5.4
configparser==5.2.0
constantly==23.10.4
crcmod==1.7
cryptography==37.0.2
cvxpy==1.3.2
cycler==0.11.0
Deprecated==1.2.13
docker-pycreds==0.4.0
docopt==0.6.2
ecos==2.0.12
einops==0.6.1
fasteners==0.17.3
filelock==3.4.2
flwr==0.17.0
fonttools==4.29.1
frozenlist==1.3.3
fsspec==2024.3.1
furl==2.1.3
gcs-oauth2-boto-plugin==3.0
gitdb==4.0.9
GitPython==3.1.26
google==2.0.3
google-api-core==2.4.0
google-apitools==0.5.32
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-reauth==0.1.1
googleapis-common-protos==1.54.0
gpustat==0.6.0
greenlet==3.0.3
grpcio==1.43.0
gsutil==5.10
gym==0.23.1
gym-notices==0.0.6
hiredis==2.0.0
httplib2==0.20.4
huggingface-hub==0.21.4
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
incremental==22.10.0
Jinja2==3.1.3
joblib==1.1.0
jsonschema==4.4.0
kiwisolver==1.3.2
learn2learn==0.1.7
-e git+https://github.com/hustvl/Vim.git@6143d07b3dd31f904c63840a19e22d95d1124493#egg=mamba_ssm&subdirectory=mamba-1p1p1
Markdown==3.3.6
MarkupSafe==2.1.5
matplotlib==3.5.1
monotonic==1.6
msgpack==1.0.3
multidict==5.2.0
mypy-extensions==0.4.3
ninja==1.10.2.3
numpy==1.21.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-ml-py3==7.352.0
oauth2client==4.1.3
oauthlib==3.1.1
opencensus==0.8.0
opencensus-context==0.1.2
orderedmultidict==1.0.1
osqp==0.6.3
packaging==24.0
pathspec==0.9.0
pathtools==0.1.2
Pillow==9.0.0
platformdirs==4.2.0
prometheus-client==0.12.0
promise==2.3
protobuf==3.19.3
psutil==5.9.0
py-spy==0.3.11
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
pyOpenSSL==22.0.0
pyparsing==3.0.6
pyrsistent==0.18.1
python-dateutil==2.8.2
pytoolconfig==1.3.1
pyu2f==0.1.5
PyYAML==6.0
qdldl==0.1.7.post0
qpth==0.0.16
ray==1.4.1
redis==4.1.1
redo==2.0.4
regex==2023.12.25
requests==2.27.1
requests-oauthlib==1.3.0
retry-decorator==1.1.1
rope==1.13.0
rsa==4.7.2
safetensors==0.4.2
scikit-learn==1.0.2
scipy==1.7.3
scs==3.2.3
sentry-sdk==1.5.2
shortuuid==1.0.8
simplejson==3.19.2
six==1.16.0
sklearn==0.0
smmap==5.0.0
soupsieve==2.3.1
SQLAlchemy==2.0.29
subprocess32==3.5.4
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.0.0
timm==0.9.16
tokenizers==0.15.2
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.62.3
transformers==4.39.1
triton==2.2.0
Twisted==24.3.0
typed-ast==1.5.1
typing_extensions==4.10.0
urllib3==1.26.8
wandb==0.12.10
webdataset==0.1.87
Werkzeug==2.0.2
wrapt==1.13.3
yarl==1.7.2
yaspin==2.1.0
zipp==3.7.0
zope.interface==6.2
